+++ NEWS.md
@@ -1,3 +1,4 @@
+-   species 0.1.0.9005: raw taxonomy script ex #4 [\@mithilaunkule](https://github.com/mithilaunkule)
 -   species 0.1.0.9004: mv appendices data_prep
 -   species 0.1.0.9003: new structure closer to R compendium
 -   species 0.1.0.9002: update links